Resolving multi-Yeslick issues
Dynaheir has crossmod content with the Pixel/Liam/Lava version of Yeslick from SHS, not with Smiling Imp's Yeslick. This would result in a
failure to install if Imp's Yeslick and Lava's Dynaheir were installed, or the banters not being installed with Lava's Dynaheir and the
SHS version of Yeslick. h/t to Lerraunt on Discord for the bug report.